Embryo_Tracker: working (hf) BlobMorphology
... based on inertia moments.
We assume we get an ellipsoid, with centroid given as spot center, and
we compute the inertia moments for axes x, y and z.
We put them in the moment of intertia tensor, which we diagonalize,
using Jama. The eignvalues are used to retrieve the ellipsoid semiaxes
length, and the eigenvectors are used to get the ellipsoid orientation.

A test main() method is provided, however I found myself unable to
generate ellipsoid with main axis not belonging to the XY plane.

+++ src-plugins/Embryo_Tracker/fiji/plugin/nperry/Spot.java
@@ -115,12 +115,12 @@ public class Spot {
 	}
 	
 	/**
-	 * Returns the {@link Double} value mapped to this {@link Feature}.
+	 * Returns the value mapped to this {@link Feature}.
 	 * @param feature The {@link Feature} to retrieve the stored value for.
-	 * @return The {@link Double} value corresponding to this {@link Feature}. 
+	 * @return The value corresponding to this {@link Feature}. 
 	 */
-	public double getFeature(Feature feature) {
-		return this.features.get(feature).doubleValue();
+	public float getFeature(Feature feature) {
+		return this.features.get(feature);
 	}
 	
 	/**
